requirements.txt
# requirements.txt
#
# Runtime dependencies for the adaptive learning path application.
# Use 'pip install -r requirements.txt' to install.
# For development, also install requirements-dev.txt

fastapi==0.111.0
uvicorn[standard]==0.29.0
neo4j==5.20.0
python-dotenv==1.0.1
pydantic==2.7.1
numpy==1.26.4
